以前のソースファイルを含めたoldディレクトリを削除、mainファイルにTOP画面生成・表示を行う暫定処理を追加
