--- eztaxcraft_app/old/app/eztc_app_main.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-
-"""
-アプリケーションのエントリーポイント
-
-Author: Shimizu
-Date: 2024/2/11
-"""
