--- eztaxcraft_app/old/app/eztc_window/eztc_login_windows.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-
-"""
-アプリのログイン画面クラス定義
-
-Author: Shimizu
-Date: 2024/2/11
-"""
