--- eztaxcraft_app/old/app/eztc_window/eztc_user_registration_windows.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-
-"""
-アプリのユーザー登録画面クラス定義
-
-Author: Shimizu
-Date: 2024/2/11
-"""
