--- eztaxcraft_app/old/app/eztc_window/eztc_window_event.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-
-"""
-アプリの画面イベントクラス定義
-
-Author: Shimizu
-Date: 2024/2/11
-"""
