--- eztaxcraft_app/old/app/eztc_window/eztc_window_manager.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-
-"""
-アプリの画面管理クラス定義
-
-Author: Shimizu
-Date: 2024/2/11
-"""
