--- eztaxcraft_app/old/common/file_control/csv_file_control.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-
-"""
-CSVファイル制御クラス定義
-
-Author: Shimizu
-Date: 2024/2/11
-"""
