--- eztaxcraft_app/old/common/window/window_control.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-
-"""
-画面制御クラス定義
-
-Author: Shimizu
-Date: 2024/2/11
-"""
